+++ src/Interpretation/InterpretEntityMetadata.fs
@@ -18,6 +18,7 @@ let typeConv = function
   | XrmAttributeType.Integer
   | XrmAttributeType.Double
   | XrmAttributeType.BigInt
+  | XrmAttributeType.Decimal
   | XrmAttributeType.Money
   | XrmAttributeType.Picklist
   | XrmAttributeType.State
@@ -26,7 +27,7 @@ let typeConv = function
 
 let interpretNormalAttribute aType (options:OptionSet option)  =
   match aType with
-  | XrmAttributeType.Money -> TsType.Number, SpecialType.Money
+  | XrmAttributeType.Money                -> TsType.Number, SpecialType.Money
     
   | XrmAttributeType.MultiSelectPicklist  -> TsType.EnumRef options.Value.name, SpecialType.MultiSelectOptionSet
   | XrmAttributeType.Picklist
@@ -40,9 +41,7 @@ let interpretNormalAttribute aType (options:OptionSet option)  =
         
   | XrmAttributeType.Uniqueidentifier     -> TsType.String, SpecialType.Guid
 
-  | XrmAttributeType.Decimal              -> typeConv aType, SpecialType.Decimal
   | _                                     -> typeConv aType, SpecialType.Default
-
 
 let interpretAttribute (nameMap: Map<string, EntityInfo>) labelMapping (a: AttributeMetadata) =
   let aType = XrmAttributeType.fromDisplayName a.AttributeTypeName
@@ -78,82 +77,21 @@ let interpretAttribute (nameMap: Map<string, EntityInfo>) labelMapping (a: Attri
     specialType = sType
     colType = aType
     targetEntitySets = targetEntitySets
-    readable = a.IsValidForRead.GetValueOrDefault(false)
-    createable = a.IsValidForCreate.GetValueOrDefault(false)
-    updateable = a.IsValidForUpdate.GetValueOrDefault(false)
+    readable = a.IsValidForRead.GetValueOrDefault false
+    createable = a.IsValidForCreate.GetValueOrDefault false
+    updateable = a.IsValidForUpdate.GetValueOrDefault false
     displayName = getLabel a.DisplayName
   }
 
-let sanitizeNavigationProptertyName string =
-    if string = null then "navigationPropertyNameNotDefined"
-    else string
 
-let interpretRelationship (nameMap: Map<string, EntityInfo>) referencing (attributes: XrmAttribute list) (rel: OneToManyRelationshipMetadata) =
-  let rLogical =
-    if referencing then rel.ReferencedEntity
-    else rel.ReferencingEntity
-    
-  Map.tryFind rLogical nameMap
-  ?|> fun eInfo ->
-    let setNames =
-      match eInfo.EntitySetName = "owners" with
-      | false -> [|eInfo.SchemaName,eInfo.EntitySetName|]
-      | true -> [|"Team","teams";"SystemUser","systemusers"|]
-    
-    let name =
-      match rel.ReferencedEntity = rel.ReferencingEntity with
-      | false -> rel.SchemaName
-      | true  ->
-        match referencing with
-        | true  -> sprintf "Referencing%s" rel.SchemaName
-        | false -> sprintf "Referenced%s" rel.SchemaName
-
-    setNames
-    |> Array.map (fun (schema,setName) ->
-      let xRel = 
-        { XrmRelationship.schemaName = name
-          attributeName = 
-            if referencing then rel.ReferencingAttribute 
-            else rel.ReferencedAttribute
-          displayName = eInfo.DisplayName
-          relType = if referencing then RelType.ManyToOne else RelType.OneToMany
-          navProp = 
-            if referencing then rel.ReferencingEntityNavigationPropertyName
-            else rel.ReferencedEntityNavigationPropertyName
-            |> sanitizeNavigationProptertyName
-          referencing = referencing
-          relatedSetName = setName
-          relatedSchemaName = schema 
-        }
-
-      eInfo.SchemaName, xRel)
-
-
-let interpretM2MRelationship (nameMap: Map<string, EntityInfo>) logicalName (rel: ManyToManyRelationshipMetadata) =
-  let rLogical =
-    match logicalName = rel.Entity2LogicalName with
-    | true  -> rel.Entity1LogicalName
-    | false -> rel.Entity2LogicalName
-    
-  Map.tryFind rLogical nameMap
-  ?|> fun eInfo ->
-      
-    let xRel = 
-      { XrmRelationship.schemaName = rel.SchemaName 
-        attributeName = rel.SchemaName
-        displayName = eInfo.DisplayName
-        relType = RelType.ManyToMany
-        navProp = 
-          if logicalName = rel.Entity2LogicalName then rel.Entity1NavigationPropertyName
-          else rel.Entity2NavigationPropertyName
-          |> sanitizeNavigationProptertyName
-        referencing = false
-        relatedSetName = eInfo.EntitySetName
-        relatedSchemaName = eInfo.SchemaName
-      }
-    
-    eInfo.SchemaName, xRel
-
+let resolveRelatedEntities (nameMap: Map<string, EntityInfo>) (logicalName: string) : EntityInfo list =
+  match Map.tryFind logicalName nameMap with
+  | Some eInfo when eInfo.EntitySetName = "owners" ->
+      let displayName k fallback = nameMap |> Map.tryFind k |> Option.map (fun e -> e.DisplayName) |> Option.defaultValue fallback
+      [ { SchemaName = "Team";       EntitySetName = "teams";       DisplayName = displayName "team"       "Team" }
+        { SchemaName = "SystemUser"; EntitySetName = "systemusers"; DisplayName = displayName "systemuser" "User" } ]
+  | Some eInfo -> [ eInfo ]
+  | None       -> []
 
 let interpretEntity (nameMap: Map<string, EntityInfo>) labelMapping (metadata:EntityMetadata) =
   if isNull metadata.Attributes then failwith "No attributes found!"
@@ -173,44 +111,16 @@ let interpretEntity (nameMap: Map<string, EntityInfo>) labelMapping (metadata:En
     |> Seq.choose id 
     |> Seq.distinctBy (fun x -> x.name) 
     |> Seq.toList
-    
 
-  let handleOneToMany referencing = function
-    | null  -> Array.empty
-    | x     -> 
-      x 
-      |> Array.choose (interpretRelationship nameMap referencing attributes)
-      |> Array.concat
-    
-  let handleManyToMany logicalName = function
-    | null  -> Array.empty
-    | x     -> x |> Array.choose (interpretM2MRelationship nameMap logicalName)
-
-
-  let relatedEntities, relationships = 
-    [ metadata.OneToManyRelationships  |> handleOneToMany false 
-      metadata.ManyToOneRelationships  |> handleOneToMany true 
-      metadata.ManyToManyRelationships |> handleManyToMany metadata.LogicalName 
-    ] |> Array.concat
-      |> List.ofArray
-      |> List.unzip
-
-  let relatedEntities = 
-    relatedEntities 
-    |> Set.ofList 
-    |> Set.remove metadata.SchemaName 
-    |> Set.toList
-
-  { XrmEntity.typecode = metadata.ObjectTypeCode.GetValueOrDefault()
-    schemaName = metadata.SchemaName
+  { XrmEntity.schemaName = metadata.SchemaName
     logicalName = metadata.LogicalName
-    entitySetName = metadata.EntitySetName |> Utility.stringToOption
-    idAttribute = metadata.PrimaryIdAttribute
-    isIntersect = metadata.IsIntersect.GetValueOrDefault(true)
+    setName = metadata.EntitySetName
+    idAttribute = attributes |> List.find (fun a -> a.logicalName = metadata.PrimaryIdAttribute)
     attributes = attributes
     optionSets = optionSets
-    relatedEntities = relatedEntities
-    allRelationships = relationships
+    oneToManyRelationships = metadata.OneToManyRelationships |> List.ofArray
+    manyToOneRelationships = metadata.ManyToOneRelationships |> List.ofArray
+    manyToManyRelationships = metadata.ManyToManyRelationships |> List.ofArray
     displayName = getLabel metadata.DisplayName
-    availableRelationships = List.empty // old
+    isIntersect = metadata.IsIntersect.GetValueOrDefault false
   }